% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll dice}
\usage{
roll(dice_string, verbose = FALSE, seed = NULL)
}
\arguments{
\item{dice_string}{A Dungeons & Dragons style dice string, e.g. "2d4" to roll
two four-sided dice and add the result}

\item{verbose}{Show the details of the roll in the console?}

\item{seed}{(optional) Random number seed to use}
}
\value{
A number representing the result of the die roll
}
\description{
Takes a Dungeons & Dragons style dice string, parses it, and
rolls it
}
\examples{
roll("d20")
roll("2d8")
roll("2d4+1")
roll("d6*3", seed =  123)
roll("3d6", verbose = TRUE)
}
